Add hoved panel for fieldsSaved
